backend/requirements.txt
# FastAPI & server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Config & env
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0

# Auth (JWT + bcrypt)
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# File / Excel (for customer upload, later)
openpyxl>=3.1.2

# HTTP client (for AI API, later)
httpx>=0.26.0

# Scheduler (recurring send)
APScheduler>=3.10.0
